fix: 프로필 이미지가 없는 경우 NPE 뜨던거 수정

+++ discordit/src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -109,7 +109,9 @@ public class BasicUserService implements UserService {
         User user = findUserAndHandleConflict(dto.userId());
         userRepository.delete(user);
         readStatusRepository.deleteAllByUser(user);
-        binaryContentRepository.delete(user.getProfileImage());
+        if (user.getProfileImage() != null) {
+            binaryContentRepository.delete(user.getProfileImage());
+        }
     }
 
     @Override
@@ -117,7 +119,9 @@ public class BasicUserService implements UserService {
         User user = findUserAndHandleConflict(uuid);
         userRepository.delete(user);
         readStatusRepository.deleteAllByUser(user);
-        binaryContentRepository.delete(user.getProfileImage());
+        if (user.getProfileImage() != null) {
+            binaryContentRepository.delete(user.getProfileImage());
+        }
     }
 
     private User findUserAndHandleConflict(UUID uuid) {
